Move configuration to a separate module
