+++ scripts/kstest-runner
@@ -26,7 +26,7 @@ import sys
 import argparse
 
 from launcher.lib.virtual_controller import VirtualManager
-from launcher.lib.configuration import VirtualConfiguration
+from launcher.lib.confs.configuration import VirtualConfiguration
 
 
 def create_parser():
