+++ scripts/launcher/lib/confs/configuration.py
