+++ scripts/launcher/lib/temp_manager.py
@@ -27,7 +27,7 @@ import shutil
 from contextlib import AbstractContextManager
 from tempfile import mkdtemp
 from glob import glob
-from .configuration import KeepLevel
+from lib.confs.configuration import KeepLevel
 
 
 class TempManager(AbstractContextManager):
